+++ os_doc_tools/commands.py
@@ -20,6 +20,8 @@ import sys
 import os_doc_tools
 from os_doc_tools.common import check_output   # noqa
 
+DEVNULL = open(os.devnull, 'wb')
+
 
 def use_help_flag(os_command):
     """Use --help flag (instead of help keyword)
@@ -63,7 +65,7 @@ def generate_heading(os_command, api_name, title, os_file):
             print("Command %s not found, aborting." % os_command)
             sys.exit(1)
     # Extract version from "swift 0.3"
-    version = version.strip().rpartition(' ')[2]
+    version = version.splitlines()[-1].strip().rpartition(' ')[2]
 
     print("Documenting '%s help (version %s)'" % (os_command, version))
 
@@ -107,6 +109,16 @@ def generate_heading(os_command, api_name, title, os_file):
     <section xml:id=\"%(os_command)sclient_command_usage\">
        <title>%(os_command)s usage</title>\n"""
 
+    if os_command == "keystone":
+        header_deprecation = """
+    <warning>
+        <para>The %(os_command)s CLI is deprecated in favor of
+            python-openstackclient. For a Python library, continue using
+            python-%(os_command)sclient.</para>
+    </warning>\n"""
+    else:
+        header_deprecation = None
+
     format_dict = {
         "os_command": os_command,
         "api_name": api_name,
@@ -115,6 +127,8 @@ def generate_heading(os_command, api_name, title, os_file):
         "help_str": help_str
     }
     os_file.write(header1 % format_dict)
+    if header_deprecation:
+        os_file.write(header_deprecation % format_dict)
     os_file.write(header2 % format_dict)
     os_file.write(header3 % format_dict)
 
@@ -264,7 +278,8 @@ def generate_command(os_command, os_file):
     :param os_file:    open filehandle for output of DocBook file
     """
 
-    help_lines = check_output([os_command, "--help"]).split('\n')
+    help_lines = check_output([os_command, "--help"],
+                              stderr=DEVNULL).split('\n')
 
     ignore_next_lines = False
     next_line_screen = True
@@ -384,7 +399,7 @@ def generate_subcommand(os_command, os_subcommand, os_file, extra_params,
     else:
         args.append("help")
         args.append(os_subcommand)
-    help_lines = check_output(args).split('\n')
+    help_lines = check_output(args, stderr=DEVNULL).split('\n')
 
     os_subcommandid = os_subcommand.replace(' ', '_')
     os_file.write("    <section xml:id=\"%sclient_subcommand_%s%s\">\n"
@@ -507,7 +522,8 @@ def get_openstack_subcommands(commands):
     """Get all subcommands of 'openstack' without using bashcompletion."""
     subcommands = []
     for command in commands:
-        output = check_output(["openstack", "help", command])
+        output = check_output(["openstack", "--os-auth-type", "token",
+                               "help", command], stderr=DEVNULL)
         for line in output.split("\n"):
             if line.strip().startswith(command):
                 subcommands.append(line.strip())
@@ -583,12 +599,13 @@ def document_single_project(os_command, output_dir):
         title = "OpenStack client"
         # Does not know about bash-completion yet, need to specify
         # commands manually and to fetch subcommands automatically
-        commands = ["aggregate", "backup", "compute", "console", "container",
+        commands = ["aggregate", "availability", "backup", "catalog",
+                    "command", "compute", "console", "container",
                     "ec2", "endpoint", "extension", "flavor", "host",
                     "hypervisor", "image", "ip", "keypair", "limits", "module",
                     "network", "object", "project", "quota", "role",
                     "security", "server", "service", "snapshot", "token",
-                    "user", "volume"]
+                    "usage", "user", "volume"]
         subcommands = get_openstack_subcommands(commands)
     else:
         print("'%s' command not yet handled" % os_command)
@@ -608,8 +625,12 @@ def document_single_project(os_command, output_dir):
     <section xml:id=\"glance_cli_v1\">
        <title>Image Service API v1 commands</title>\n""")
 
-    generate_subcommands(os_command, out_file, blacklist,
-                         subcommands, None, "", "")
+    if os_command == 'openstack':
+        generate_subcommands(os_command, out_file, blacklist,
+                             subcommands, ["--os-auth-type", "token"], "", "")
+    else:
+        generate_subcommands(os_command, out_file, blacklist,
+                             subcommands, None, "", "")
 
     if os_command == 'cinder':
         out_file.write("    </section>\n")